chore: Move additional hosting properties to IWpfXamlRootHost
