+++ src/Uno.UI.Runtime.Skia.Wpf/Extensions/WpfApplicationViewExtension.cs
