+++ src/Uno.UI.Runtime.Skia.Wpf/Extensions/WpfCoreWindowExtension.cs
