+++ src/Uno.UI.Runtime.Skia.Wpf/Extensions/WpfDisplayInformationExtension.cs
